F18 : display home

+++ app/src/main/java/c0d3/vitreen/app/models/mini/AdvertMini.kt
@@ -4,5 +4,5 @@ data class AdvertMini(
         val id:String,
         val title: String,
         val description: String,
-        val price: Float
+        val price: Long
 )
